fix fiat ludum section bleeding into next section
Top section would bleed into next section if the window was resized to
have a small height. Made the image resizing based on viewport height to
slightly mitigate the problem (there's still bleeding when the window is
resized to be very short, but most people don't have a screen with those
dimensions).